+++ examples/basic_table_button.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+require 'glimmer-dsl-libui'
+
+include Glimmer
+
+data = [
+  %w[cat meow delete],
+  %w[dog woof delete],
+  %w[chicken cock-a-doodle-doo delete],
+  %w[hourse neigh delete],
+  %w[cow moo delete]
+]
+
+window('Animal sounds', 300, 200) {
+  horizontal_box {
+    table {
+      text_column('Animal')
+      text_column('Description')
+      button_column('Action') {
+#         on_clicked do |row|
+#           data.delete(row)
+#         end
+      }
+
+      cell_rows data
+    }
+  }
+  
+  on_closing do
+    puts 'Bye Bye'
+  end
+}.show